EShop/requirements.txt
Django==2.2.4
django-jet==1.0.8
djangorestframework==3.10.3
psycopg2-binary==2.8.4
Pillow==7.0.0
django-phone-field==1.8.0
django_extensions==2.2.8
django-rest-registration==0.5.3
#social_django==3.1.0
social-auth-app-django
django-mptt==0.11.0
django-rest-registration==0.5.3
django-extensions==2.2.8
djangorestframework-recursive==0.1.2
redis==3.4.1
celery==4.4.0